% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2path-package.R
\docType{package}
\name{string2path-package}
\alias{string2path}
\alias{string2path-package}
\title{string2path: Convert a string to path}
\description{
An experimental package to use Rust crates; rusttype <https://crates.io/crates/rusttype> for
    reading and extracting TTF file, and lyon <https://crates.io/crates/lyon> to translate curves to
    flattened paths.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
\keyword{internal}
