% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\docType{package}
\name{string2path-package}
\alias{string2path}
\alias{string2path-package}
\title{string2path: Rendering Font into 'data.frame'}
\description{
Extract glyph information from font data, and translate the outline curves to flattened paths or tessellated polygons. The converted data is returned as a 'data.frame' in easy-to-plot format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yutannihilation.github.io/string2path/}
  \item \url{https://github.com/yutannihilation/string2path}
  \item Report bugs at \url{https://github.com/yutannihilation/string2path/issues}
}

}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
