% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{string2path}
\alias{string2path}
\alias{string2stroke}
\alias{string2fill}
\title{Convert strings to paths}
\usage{
string2path(text, font_file, tolerance = 0.001)

string2stroke(text, font_file, tolerance = 0.001, line_width = 0.1)

string2fill(text, font_file, tolerance = 0.001)
}
\arguments{
\item{text}{A text to convert to paths.}

\item{font_file}{A path to TTF or OTF file.}

\item{tolerance}{Maximum distance allowed between the curve and its approximation. For more details,
please refer to \href{https://docs.rs/lyon/0.16.2/lyon/index.html#what-is-the-tolerance-variable-in-these-examples}{lyon's doc}.}

\item{line_width}{Line width of strokes.}

\item{result_type}{Result types.}
}
\description{
Convert strings to paths
}
